megacommit: binary writing update, v1 binary reading, many bug fixes
Binary writing needs more testing but seems to be working well in a few tests for now. It won't produce an exact match to the original file (there seems to be some weird ordering going on in the original file) and is slightly larger, but otherwise still seems to work.

V1 binary support and caching is supported thanks to LukeFZ. Some older json catalog support is supported thanks to C-Sharper.

Most likely, many things are still broken (especially with writing.) V1 binaries current corrupt on write which will be fixed in a later commit.

Co-authored-by: LukeFZ <17146677+LukeFZ@users.noreply.github.com>
Co-authored-by: C-SHARPER <c.sharper.alt@gmail.com>